#307 批量处理过滤作业

+++ saturn-console-web/src/pages/job_list/job_overview.vue
@@ -94,7 +94,7 @@
                                     <el-tooltip content="复制" placement="top">
                                         <el-button type="text" @click="handleCopy(scope.row)"><i class="fa fa-clone"></i></el-button>
                                     </el-tooltip>
-                                    <el-tooltip content="删除" placement="top">
+                                    <el-tooltip content="删除" placement="top" v-if="scope.row.status === 'STOPPED'">
                                         <el-button type="text" icon="el-icon-delete" @click="handleDelete(scope.row)"></el-button>
                                     </el-tooltip>
                                 </template>
@@ -217,37 +217,53 @@ export default {
     },
     batchEnabled() {
       this.batchOperation('启用', (arr) => {
-        const params = {
-          jobNames: arr.join(','),
-        };
-        this.handleBatchActive(params, arr, true);
+        const stopedJob = this.getStoptedJobArray(arr);
+        if (stopedJob.length === 0) {
+          this.$message.errorMessage('没有可以启用的作业,请重新勾选!');
+        } else {
+          const params = {
+            jobNames: this.getJobNameArray(stopedJob).join(','),
+          };
+          this.handleBatchActive(params, this.getJobNameArray(stopedJob), true);
+        }
       });
     },
     batchDisabled() {
       this.batchOperation('禁用', (arr) => {
-        const params = {
-          jobNames: arr.join(','),
-        };
-        this.handleBatchActive(params, arr, false);
+        const unstopedJob = this.getUnstoptedJobArray(arr);
+        if (unstopedJob.length === 0) {
+          this.$message.errorMessage('没有可以禁用的作业,请重新勾选!');
+        } else {
+          const params = {
+            jobNames: this.getJobNameArray(unstopedJob).join(','),
+          };
+          this.handleBatchActive(params, this.getJobNameArray(unstopedJob), false);
+        }
       });
     },
     batchDelete() {
       this.batchOperation('删除', (arr) => {
-        const params = {
-          jobNames: arr.join(','),
-        };
-        this.$message.confirmMessage(`确认删除作业 ${params.jobNames} 吗?`, () => {
-          this.$http.delete(`/console/namespaces/${this.domainName}/jobs`, params).then(() => {
-            this.getJobList();
-            this.$message.successNotify('批量删除作业操作成功');
-          })
-          .catch(() => { this.$http.buildErrorHandler('批量删除作业请求失败！'); });
-        });
+        const stopedJob = this.getStoptedJobArray(arr);
+        if (stopedJob.length === 0) {
+          this.$message.errorMessage('没有可以删除的作业,请重新勾选!');
+        } else {
+          const params = {
+            jobNames: this.getJobNameArray(stopedJob).join(','),
+          };
+          this.$message.confirmMessage(`确认删除作业 ${params.jobNames} 吗?`, () => {
+            this.$http.delete(`/console/namespaces/${this.domainName}/jobs`, params).then(() => {
+              this.getJobList();
+              this.$message.successNotify('批量删除作业操作成功');
+            })
+            .catch(() => { this.$http.buildErrorHandler('批量删除作业请求失败！'); });
+          });
+        }
       });
     },
     batchPriority() {
       this.batchOperation('优先Executors', (arr) => {
-        this.jobNamesArray = JSON.parse(JSON.stringify(arr));
+        const jobNames = this.getJobNameArray(arr);
+        this.jobNamesArray = JSON.parse(JSON.stringify(jobNames));
         this.isBatchPriorityVisible = true;
       });
     },
@@ -265,7 +281,7 @@ export default {
       } else {
         const selectedJobNameArray = [];
         this.multipleSelection.forEach((element) => {
-          selectedJobNameArray.push(element.jobName);
+          selectedJobNameArray.push(element);
         });
         callback(selectedJobNameArray);
       }
@@ -415,6 +431,31 @@ export default {
         this.getJobList();
       })
       .catch(() => { this.$http.buildErrorHandler(`${reqUrl}请求失败！`); });
+    },
+    getJobNameArray(arr) {
+      const jobNameArray = [];
+      arr.forEach((ele) => {
+        jobNameArray.push(ele.jobName);
+      });
+      return jobNameArray;
+    },
+    getUnstoptedJobArray(arr) {
+      const resultArr = [];
+      arr.forEach((ele) => {
+        if (ele.status !== 'STOPPED') {
+          resultArr.push(ele);
+        }
+      });
+      return resultArr;
+    },
+    getStoptedJobArray(arr) {
+      const resultArr = [];
+      arr.forEach((ele) => {
+        if (ele.status === 'STOPPED') {
+          resultArr.push(ele);
+        }
+      });
+      return resultArr;
     },
     getJobList() {
       this.loading = true;
